refactor dpm campaign approval routes
